Rubber-stamped by Mark Rowe.
Fix for loop to use an size_t instead of unsigned and some spacing
style fixes.

* dom/Node.cpp:
(WebCore::eventHasListeners):



Canonical link: https://commits.webkit.org/41838@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@50405 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/dom/Node.cpp
@@ -2446,12 +2446,12 @@ bool Node::dispatchEvent(PassRefPtr<Event> prpEvent)
     return dispatchGenericEvent(event.release());
 }
 
-static bool eventHasListeners(const AtomicString& eventType, Node* node, Vector< RefPtr< ContainerNode > >& ancestors)
+static bool eventHasListeners(const AtomicString& eventType, Node* node, Vector<RefPtr<ContainerNode> >& ancestors)
 {
     if (node->hasEventListeners(eventType))
         return true;
 
-    for (unsigned i = 0; i < ancestors.size(); i++) {
+    for (size_t i = 0; i < ancestors.size(); i++) {
         ContainerNode* ancestor = ancestors[i].get();
         if (ancestor->hasEventListeners(eventType))
             return true;
